erp-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-barcode
pandas
openpyxl
yagmail
aiofiles
email-validator
reportlab
python-multipart